+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## 0.5.0.dev
+## 0.5.0
 
 ### Sentence corpus creation
 
@@ -12,12 +12,13 @@
 
 - Add consolidation logic to group quotes that are sequential in both original and reuse texts
     - Quote consolidation is enabled by default but configurable in both app and command line script
+- Now supports multiple original sentence corpora as input
+    - `remarx-find-quotes` script accepts one or more CSV files or directories of CSVs for original corpora; uses default original corpus location when no original path is specified
 
 ### Scripts
 
 - Add `xquery` directory with instructions and XQuery files for investigating
     & reporting on continuing paragraph and footnote content in MEGA TEI
-- Quote finder CLI script now accepts either a single corpus CSV or a directory of corpora, and defaults to `~/remarx-data/corpora/original` when no original path is provided
 
 ## 0.4.0
 
@@ -39,7 +40,7 @@
 
 ### Quotation detection
 
-- Add the `find_quotes.py` CLI script to run the quote finder, with a `--benchmark` flag to collect performance metrics
+- Add `remarx-find-quotes` script to run the quote finder, with a `--benchmark` option to collect performance metrics
 
 ### Application